+++ packages/i18n-calypso/src/index.ts
@@ -10,7 +10,6 @@ export default i18n;
 
 // Export the default instance's properties and bound methods for convenience
 // These should be deprecated eventually, exposing only the default `i18n` instance
-export const geolocateCurrencySymbol = i18n.geolocateCurrencySymbol.bind( i18n );
 export const translate = i18n.translate.bind( i18n );
 export const configure = i18n.configure.bind( i18n );
 export const setLocale = i18n.setLocale.bind( i18n );