+++ packages/i18n-calypso/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Breaking: Emitter object replaced with subscriber
   - Before: `i18n.on( 'change', callback ); i18n.off( 'change', callback );`
   - After: `const unsubscribe = i18n.subscribe( callback ); unsubscribe();`
+- Breaking: Removed `geolocateCurrencySymbol` function. Country code assignment is handled in the `format-currency` package `setGeoLocation` function.
 
 ## 7.5.0